Removing pytest qt wrapper
